Enable Word Auto-Pagination Using Group Variable

This function uses the applies the functionality flextable::keep_with_next() by automatically building the row indices using some grouping variable. Each group identified by the variable (i.e. when the value of the variable changes) will be set as a "keep_with_next" group in Word. Using this functionality, Word will attempt not to break that group across pages, enabling smoother pagination without having to do specific calculations of page breaks.

Usage

clin_auto_page(x, group_var)

Arguments

x

A clintable object

group_var

A string containing a variable name of the input dataset used to calculate groups

Value

A clintable object

Examples

clintable(mtcars) |>
  clin_auto_page("gear")

mpg

cyl

disp

hp

drat

wt

qsec

vs

am

carb

21.0

6

160.0

110

3.90

2.620

16.46

0

1

4

21.0

6

160.0

110

3.90

2.875

17.02

0

1

4

22.8

4

108.0

93

3.85

2.320

18.61

1

1

1

21.4

6

258.0

110

3.08

3.215

19.44

1

0

1

18.7

8

360.0

175

3.15

3.440

17.02

0

0

2

18.1

6

225.0

105

2.76

3.460

20.22

1

0

1

14.3

8

360.0

245

3.21

3.570

15.84

0

0

4

24.4

4

146.7

62

3.69

3.190

20.00

1

0

2

22.8

4

140.8

95

3.92

3.150

22.90

1

0

2

19.2

6

167.6

123

3.92

3.440

18.30

1

0

4

17.8

6

167.6

123

3.92

3.440

18.90

1

0

4

16.4

8

275.8

180

3.07

4.070

17.40

0

0

3

17.3

8

275.8

180

3.07

3.730

17.60

0

0

3

15.2

8

275.8

180

3.07

3.780

18.00

0

0

3

10.4

8

472.0

205

2.93

5.250

17.98

0

0

4