Set column widths using percent

Extraction of flextable print method with special handling of clintable pages and

Usage

clin_col_widths(x, ...)

Arguments

x

A clintable object

...

Named parameters where the names are columns in the flextable and the values are decimals representing the percent of total width of the table

Value

A clintable object

Examples

ct <- clintable(mtcars)

ct <- clin_alt_pages(
  ct,
  key_cols = c("mpg", "cyl", "hp"),
  col_groups = list(
    c("disp", "drat", "wt"),
    c("qsec", "vs", "am"),
    c("gear", "carb")
  )
) |>
  clin_col_widths(mpg = .2, cyl = .2, disp = .15, vs = .15)

print(ct)
#> NOTE: Alternating pages were set, but no selection for row wise pagination was configured Defaulting to 20 rows per page.
#> <div id="page1" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-0970c8ec{}.cl-096907ba{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-096bb4ce{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-096bb4e2{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-096bd170{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-096bd184{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-096bd185{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-096bd186{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-0970c8ec'><thead><tr style="overflow-wrap:break-word;"><th class="cl-096bd170"><p class="cl-096bb4ce"><span class="cl-096907ba">mpg</span></p></th><th class="cl-096bd170"><p class="cl-096bb4ce"><span class="cl-096907ba">cyl</span></p></th><th class="cl-096bd184"><p class="cl-096bb4ce"><span class="cl-096907ba">hp</span></p></th><th class="cl-096bd184"><p class="cl-096bb4ce"><span class="cl-096907ba">disp</span></p></th><th class="cl-096bd184"><p class="cl-096bb4ce"><span class="cl-096907ba">drat</span></p></th><th class="cl-096bd184"><p class="cl-096bb4ce"><span class="cl-096907ba">wt</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">21.0</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">6</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">110</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">160.0</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.90</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">2.620</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">21.0</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">6</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">110</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">160.0</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.90</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">2.875</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">22.8</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">4</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">93</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">108.0</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.85</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">2.320</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">21.4</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">6</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">110</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">258.0</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.08</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.215</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">18.7</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">8</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">175</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">360.0</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.15</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">18.1</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">6</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">105</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">225.0</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">2.76</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.460</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">14.3</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">8</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">245</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">360.0</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.21</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.570</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">24.4</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">4</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">62</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">146.7</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.69</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.190</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">22.8</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">4</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">95</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">140.8</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.92</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.150</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">19.2</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">6</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">123</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">167.6</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.92</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">17.8</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">6</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">123</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">167.6</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.92</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">16.4</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">8</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">180</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">275.8</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.07</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">4.070</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">17.3</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">8</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">180</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">275.8</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.07</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.730</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">15.2</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">8</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">180</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">275.8</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.07</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.780</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">10.4</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">8</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">205</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">472.0</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">2.93</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">5.250</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">10.4</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">8</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">215</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">460.0</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.00</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">5.424</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">14.7</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">8</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">230</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">440.0</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">3.23</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">5.345</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">32.4</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">4</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">66</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">78.7</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">4.08</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">2.200</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">30.4</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">4</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">52</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">75.7</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">4.93</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">1.615</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">33.9</span></p></td><td class="cl-096bd185"><p class="cl-096bb4e2"><span class="cl-096907ba">4</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">65</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">71.1</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">4.22</span></p></td><td class="cl-096bd186"><p class="cl-096bb4e2"><span class="cl-096907ba">1.835</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page2" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-097dac88{}.cl-0975cbee{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-0978a012{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-0978a01c{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-0978be30{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-0978be31{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-0978be3a{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-0978be3b{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-0978be44{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-0978be45{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-097dac88'><thead><tr style="overflow-wrap:break-word;"><th class="cl-0978be30"><p class="cl-0978a012"><span class="cl-0975cbee">mpg</span></p></th><th class="cl-0978be30"><p class="cl-0978a012"><span class="cl-0975cbee">cyl</span></p></th><th class="cl-0978be31"><p class="cl-0978a012"><span class="cl-0975cbee">hp</span></p></th><th class="cl-0978be3a"><p class="cl-0978a012"><span class="cl-0975cbee">qsec</span></p></th><th class="cl-0978be31"><p class="cl-0978a012"><span class="cl-0975cbee">vs</span></p></th><th class="cl-0978be3a"><p class="cl-0978a012"><span class="cl-0975cbee">am</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">21.0</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">6</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">110</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">16.46</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">21.0</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">6</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">110</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">17.02</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">22.8</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">4</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">93</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">18.61</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">21.4</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">6</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">110</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">19.44</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">18.7</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">8</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">175</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">17.02</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">18.1</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">6</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">105</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">20.22</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">14.3</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">8</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">245</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">15.84</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">24.4</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">4</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">62</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">20.00</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">22.8</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">4</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">95</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">22.90</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">19.2</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">6</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">123</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">18.30</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">17.8</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">6</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">123</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">18.90</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">16.4</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">8</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">180</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">17.40</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">17.3</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">8</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">180</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">17.60</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">15.2</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">8</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">180</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">18.00</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">10.4</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">8</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">205</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">17.98</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">10.4</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">8</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">215</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">17.82</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">14.7</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">8</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">230</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">17.42</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">32.4</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">4</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">66</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">19.47</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">30.4</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">4</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">52</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">18.52</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">33.9</span></p></td><td class="cl-0978be3b"><p class="cl-0978a01c"><span class="cl-0975cbee">4</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">65</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">19.90</span></p></td><td class="cl-0978be44"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td><td class="cl-0978be45"><p class="cl-0978a01c"><span class="cl-0975cbee">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page3" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-0988c0f0{}.cl-098241da{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-0984d9b8{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-0984d9c2{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-0984f5a6{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-0984f5b0{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-0984f5ba{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-0984f5bb{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-0984f5bc{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-0984f5c4{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-0988c0f0'><thead><tr style="overflow-wrap:break-word;"><th class="cl-0984f5a6"><p class="cl-0984d9b8"><span class="cl-098241da">mpg</span></p></th><th class="cl-0984f5a6"><p class="cl-0984d9b8"><span class="cl-098241da">cyl</span></p></th><th class="cl-0984f5b0"><p class="cl-0984d9b8"><span class="cl-098241da">hp</span></p></th><th class="cl-0984f5ba"><p class="cl-0984d9b8"><span class="cl-098241da">gear</span></p></th><th class="cl-0984f5ba"><p class="cl-0984d9b8"><span class="cl-098241da">carb</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">21.0</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">6</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">110</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">21.0</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">6</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">110</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">22.8</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">93</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">21.4</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">6</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">110</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">3</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">18.7</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">8</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">175</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">3</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">18.1</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">6</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">105</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">3</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">14.3</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">8</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">245</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">3</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">24.4</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">62</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">22.8</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">95</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">19.2</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">6</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">123</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">17.8</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">6</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">123</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">16.4</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">8</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">180</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">3</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">17.3</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">8</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">180</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">3</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">15.2</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">8</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">180</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">3</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">10.4</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">8</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">205</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">3</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">10.4</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">8</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">215</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">3</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">14.7</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">8</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">230</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">3</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">32.4</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">66</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">30.4</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">52</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">33.9</span></p></td><td class="cl-0984f5bb"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5bc"><p class="cl-0984d9c2"><span class="cl-098241da">65</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">4</span></p></td><td class="cl-0984f5c4"><p class="cl-0984d9c2"><span class="cl-098241da">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div>
#>   <button onclick="showPage(1)">1</button>
#>   <button onclick="showPage(2)">2</button>
#>   <button onclick="showPage(3)">3</button>
#> </div>
#> <script>
#>   function showPage(pageNum) {
#>     var pages = document.querySelectorAll('.page');
#>     pages.forEach(page => page.style.display = 'none');
#>     document.getElementById('page' + pageNum).style.display = 'block';
#>   }
#> </script>