Clintable write method

Write a clinify table out to a docx file

Usage

write_clintable(x, file, apply_defaults = TRUE)

Arguments

x

a clintable object

file

The file path to which the file should be written

apply_defaults

Apply default styles. These styles are stored in the options clinify_header_default, clinify_footer_default, and clinify_table_default respectively. Defaults to true.

Value

Invisible

Examples

if (FALSE) { # \dontrun{
ct <- clintable(mtcars)

ct <- clin_alt_pages(
  ct,
  key_cols = c('mpg', 'cyl', 'hp'),
  col_groups = list(
    c('disp', 'drat', 'wt'),
    c('qsec', 'vs', 'am'),
    c('gear', 'carb')
  )
)

write_clintable(ct, 'demo.docx')
} # }