Add page footer

This function adds a text, passed as argument, into .docx bottom colontitle (page footer). As a result, the text is repeated on each page of the document. Takes a list of lists as an input parameter, where each sublist represents a separate line (or row) of text. There is also a possibility to add a row with both left- and right-aligned text. To do so - simply pass two items to a sublist instead of one.

Usage

clin_add_titles(x, ls = NULL, ft = NULL)

clin_add_footnotes(x, ls = NULL, ft = NULL)

Arguments

x

a clintable object

ls

a list of character vectors, no more than 3 elements to a vector

ft

A flextable object to use as the header

Value

A clintable object

Examples

clintable(mtcars) |> 
  clin_add_titles(
    list(
      c("Left", "Center", "Right"),
      c("Just the middle")
    )
  ) |> 
  clin_add_footnotes(
    list(
      c(
        "Here's a footnote.", 
        format(Sys.time(), "%H:%M %A, %B %d, %Y")
      )
    )
  )