Set column widths using percent

Extraction of flextable print method with special handling of clintable pages and

Usage

clin_col_widths(x, ...)

Arguments

x

A clintable object

...

Named parameters where the names are columns in the flextable and the values are decimals representing the percent of total width of the table

Value

A clintable object

Examples

ct <- clintable(mtcars)

ct <- clin_alt_pages(
  ct,
  key_cols = c('mpg', 'cyl', 'hp'),
  col_groups = list(
    c('disp', 'drat', 'wt'),
    c('qsec', 'vs', 'am'),
    c('gear', 'carb')
  )
) |> 
clin_col_widths(mpg = .2, cyl=.2, disp=.15, vs=.15)

print(ct)
#> NOTE: Alternating pages were set, but no selection for row wise pagination was configured Defaulting to 20 rows per page.
#> <div id="page1" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-a96b32b0{}.cl-a9648ece{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-a9672fe4{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-a9672fee{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-a9674dc6{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a9674dd0{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a9674dd1{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a9674dda{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-a96b32b0'><thead><tr style="overflow-wrap:break-word;"><th class="cl-a9674dc6"><p class="cl-a9672fe4"><span class="cl-a9648ece">mpg</span></p></th><th class="cl-a9674dc6"><p class="cl-a9672fe4"><span class="cl-a9648ece">cyl</span></p></th><th class="cl-a9674dd0"><p class="cl-a9672fe4"><span class="cl-a9648ece">hp</span></p></th><th class="cl-a9674dd0"><p class="cl-a9672fe4"><span class="cl-a9648ece">disp</span></p></th><th class="cl-a9674dd0"><p class="cl-a9672fe4"><span class="cl-a9648ece">drat</span></p></th><th class="cl-a9674dd0"><p class="cl-a9672fe4"><span class="cl-a9648ece">wt</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">21.0</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">6</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">110</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">160.0</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.90</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">2.620</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">21.0</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">6</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">110</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">160.0</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.90</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">2.875</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">22.8</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">4</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">93</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">108.0</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.85</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">2.320</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">21.4</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">6</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">110</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">258.0</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.08</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.215</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">18.7</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">8</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">175</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">360.0</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.15</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">18.1</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">6</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">105</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">225.0</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">2.76</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.460</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">14.3</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">8</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">245</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">360.0</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.21</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.570</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">24.4</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">4</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">62</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">146.7</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.69</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.190</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">22.8</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">4</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">95</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">140.8</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.92</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.150</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">19.2</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">6</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">123</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">167.6</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.92</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">17.8</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">6</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">123</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">167.6</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.92</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">16.4</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">8</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">180</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">275.8</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.07</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">4.070</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">17.3</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">8</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">180</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">275.8</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.07</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.730</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">15.2</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">8</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">180</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">275.8</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.07</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.780</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">10.4</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">8</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">205</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">472.0</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">2.93</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">5.250</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">10.4</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">8</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">215</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">460.0</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.00</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">5.424</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">14.7</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">8</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">230</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">440.0</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">3.23</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">5.345</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">32.4</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">4</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">66</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">78.7</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">4.08</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">2.200</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">30.4</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">4</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">52</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">75.7</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">4.93</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">1.615</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">33.9</span></p></td><td class="cl-a9674dd1"><p class="cl-a9672fee"><span class="cl-a9648ece">4</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">65</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">71.1</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">4.22</span></p></td><td class="cl-a9674dda"><p class="cl-a9672fee"><span class="cl-a9648ece">1.835</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page2" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-a97730d8{}.cl-a970cb62{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-a9734dc4{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-a9734dd8{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-a97369bc{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a97369c6{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a97369c7{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a97369d0{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a97369d1{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a97369da{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-a97730d8'><thead><tr style="overflow-wrap:break-word;"><th class="cl-a97369bc"><p class="cl-a9734dc4"><span class="cl-a970cb62">mpg</span></p></th><th class="cl-a97369bc"><p class="cl-a9734dc4"><span class="cl-a970cb62">cyl</span></p></th><th class="cl-a97369c6"><p class="cl-a9734dc4"><span class="cl-a970cb62">hp</span></p></th><th class="cl-a97369c7"><p class="cl-a9734dc4"><span class="cl-a970cb62">qsec</span></p></th><th class="cl-a97369c6"><p class="cl-a9734dc4"><span class="cl-a970cb62">vs</span></p></th><th class="cl-a97369c7"><p class="cl-a9734dc4"><span class="cl-a970cb62">am</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">21.0</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">6</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">110</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">16.46</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">21.0</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">6</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">110</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">17.02</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">22.8</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">4</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">93</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">18.61</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">21.4</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">6</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">110</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">19.44</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">18.7</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">8</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">175</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">17.02</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">18.1</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">6</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">105</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">20.22</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">14.3</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">8</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">245</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">15.84</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">24.4</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">4</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">62</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">20.00</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">22.8</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">4</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">95</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">22.90</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">19.2</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">6</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">123</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">18.30</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">17.8</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">6</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">123</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">18.90</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">16.4</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">8</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">180</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">17.40</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">17.3</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">8</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">180</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">17.60</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">15.2</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">8</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">180</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">18.00</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">10.4</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">8</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">205</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">17.98</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">10.4</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">8</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">215</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">17.82</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">14.7</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">8</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">230</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">17.42</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">32.4</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">4</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">66</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">19.47</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">30.4</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">4</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">52</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">18.52</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">33.9</span></p></td><td class="cl-a97369d0"><p class="cl-a9734dd8"><span class="cl-a970cb62">4</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">65</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">19.90</span></p></td><td class="cl-a97369d1"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td><td class="cl-a97369da"><p class="cl-a9734dd8"><span class="cl-a970cb62">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page3" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-a9835fb6{}.cl-a97c2958{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-a97fa7ea{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-a97fa7fe{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-a97fc3c4{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a97fc3ce{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a97fc3cf{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a97fc3d0{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a97fc3d8{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-a97fc3d9{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-a9835fb6'><thead><tr style="overflow-wrap:break-word;"><th class="cl-a97fc3c4"><p class="cl-a97fa7ea"><span class="cl-a97c2958">mpg</span></p></th><th class="cl-a97fc3c4"><p class="cl-a97fa7ea"><span class="cl-a97c2958">cyl</span></p></th><th class="cl-a97fc3ce"><p class="cl-a97fa7ea"><span class="cl-a97c2958">hp</span></p></th><th class="cl-a97fc3cf"><p class="cl-a97fa7ea"><span class="cl-a97c2958">gear</span></p></th><th class="cl-a97fc3cf"><p class="cl-a97fa7ea"><span class="cl-a97c2958">carb</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">21.0</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">6</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">110</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">21.0</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">6</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">110</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">22.8</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">93</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">21.4</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">6</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">110</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">3</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">18.7</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">8</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">175</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">3</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">18.1</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">6</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">105</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">3</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">14.3</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">8</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">245</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">3</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">24.4</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">62</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">22.8</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">95</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">19.2</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">6</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">123</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">17.8</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">6</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">123</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">16.4</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">8</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">180</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">3</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">17.3</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">8</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">180</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">3</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">15.2</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">8</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">180</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">3</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">10.4</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">8</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">205</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">3</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">10.4</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">8</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">215</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">3</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">14.7</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">8</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">230</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">3</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">32.4</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">66</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">30.4</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">52</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">33.9</span></p></td><td class="cl-a97fc3d0"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d8"><p class="cl-a97fa7fe"><span class="cl-a97c2958">65</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">4</span></p></td><td class="cl-a97fc3d9"><p class="cl-a97fa7fe"><span class="cl-a97c2958">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div>
#>   <button onclick="showPage(1)">1</button>
#>   <button onclick="showPage(2)">2</button>
#>   <button onclick="showPage(3)">3</button>
#> </div>
#> <script>
#>   function showPage(pageNum) {
#>     var pages = document.querySelectorAll('.page');
#>     pages.forEach(page => page.style.display = 'none');
#>     document.getElementById('page' + pageNum).style.display = 'block';
#>   }
#> </script>