Set column widths using percent

Extraction of flextable print method with special handling of clintable pages and

Usage

clin_col_widths(x, ...)

Arguments

x

A clintable object

...

Named parameters where the names are columns in the flextable and the values are decimals representing the percent of total width of the table

Value

A clintable object

Examples

ct <- clintable(mtcars)

ct <- clin_alt_pages(
  ct,
  key_cols = c('mpg', 'cyl', 'hp'),
  col_groups = list(
    c('disp', 'drat', 'wt'),
    c('qsec', 'vs', 'am'),
    c('gear', 'carb')
  )
) |> 
clin_col_widths(mpg = .2, cyl=.2, disp=.15, vs=.15)

print(ct)
#> NOTE: Alternating pages were set, but no selection for row wise pagination was configured Defaulting to 20 rows per page.
#> <div id="page1" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-735a72c4{}.cl-73541532{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-735696c2{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-735696cc{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-7356b33c{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-7356b346{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-7356b347{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-7356b350{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-735a72c4'><thead><tr style="overflow-wrap:break-word;"><th class="cl-7356b33c"><p class="cl-735696c2"><span class="cl-73541532">mpg</span></p></th><th class="cl-7356b33c"><p class="cl-735696c2"><span class="cl-73541532">cyl</span></p></th><th class="cl-7356b346"><p class="cl-735696c2"><span class="cl-73541532">hp</span></p></th><th class="cl-7356b346"><p class="cl-735696c2"><span class="cl-73541532">disp</span></p></th><th class="cl-7356b346"><p class="cl-735696c2"><span class="cl-73541532">drat</span></p></th><th class="cl-7356b346"><p class="cl-735696c2"><span class="cl-73541532">wt</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">21.0</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">6</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">110</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">160.0</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.90</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">2.620</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">21.0</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">6</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">110</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">160.0</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.90</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">2.875</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">22.8</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">4</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">93</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">108.0</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.85</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">2.320</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">21.4</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">6</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">110</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">258.0</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.08</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.215</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">18.7</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">8</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">175</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">360.0</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.15</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">18.1</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">6</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">105</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">225.0</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">2.76</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.460</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">14.3</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">8</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">245</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">360.0</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.21</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.570</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">24.4</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">4</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">62</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">146.7</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.69</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.190</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">22.8</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">4</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">95</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">140.8</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.92</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.150</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">19.2</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">6</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">123</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">167.6</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.92</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">17.8</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">6</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">123</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">167.6</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.92</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">16.4</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">8</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">180</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">275.8</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.07</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">4.070</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">17.3</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">8</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">180</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">275.8</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.07</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.730</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">15.2</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">8</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">180</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">275.8</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.07</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.780</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">10.4</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">8</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">205</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">472.0</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">2.93</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">5.250</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">10.4</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">8</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">215</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">460.0</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.00</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">5.424</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">14.7</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">8</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">230</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">440.0</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">3.23</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">5.345</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">32.4</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">4</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">66</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">78.7</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">4.08</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">2.200</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">30.4</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">4</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">52</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">75.7</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">4.93</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">1.615</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">33.9</span></p></td><td class="cl-7356b347"><p class="cl-735696cc"><span class="cl-73541532">4</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">65</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">71.1</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">4.22</span></p></td><td class="cl-7356b350"><p class="cl-735696cc"><span class="cl-73541532">1.835</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page2" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-73660684{}.cl-735f4b3c{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-7362594e{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-73625958{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-73627366{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73627370{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73627371{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73627372{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73627373{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-7362737a{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-73660684'><thead><tr style="overflow-wrap:break-word;"><th class="cl-73627366"><p class="cl-7362594e"><span class="cl-735f4b3c">mpg</span></p></th><th class="cl-73627366"><p class="cl-7362594e"><span class="cl-735f4b3c">cyl</span></p></th><th class="cl-73627370"><p class="cl-7362594e"><span class="cl-735f4b3c">hp</span></p></th><th class="cl-73627371"><p class="cl-7362594e"><span class="cl-735f4b3c">qsec</span></p></th><th class="cl-73627370"><p class="cl-7362594e"><span class="cl-735f4b3c">vs</span></p></th><th class="cl-73627371"><p class="cl-7362594e"><span class="cl-735f4b3c">am</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">21.0</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">6</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">110</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">16.46</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">21.0</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">6</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">110</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">17.02</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">22.8</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">4</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">93</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">18.61</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">21.4</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">6</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">110</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">19.44</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">18.7</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">8</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">175</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">17.02</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">18.1</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">6</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">105</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">20.22</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">14.3</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">8</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">245</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">15.84</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">24.4</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">4</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">62</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">20.00</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">22.8</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">4</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">95</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">22.90</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">19.2</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">6</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">123</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">18.30</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">17.8</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">6</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">123</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">18.90</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">16.4</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">8</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">180</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">17.40</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">17.3</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">8</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">180</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">17.60</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">15.2</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">8</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">180</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">18.00</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">10.4</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">8</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">205</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">17.98</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">10.4</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">8</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">215</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">17.82</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">14.7</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">8</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">230</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">17.42</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">32.4</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">4</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">66</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">19.47</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">30.4</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">4</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">52</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">18.52</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">33.9</span></p></td><td class="cl-73627372"><p class="cl-73625958"><span class="cl-735f4b3c">4</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">65</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">19.90</span></p></td><td class="cl-73627373"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td><td class="cl-7362737a"><p class="cl-73625958"><span class="cl-735f4b3c">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page3" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-7370df50{}.cl-736a370e{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-736c9f58{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-736c9f62{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-736cbb82{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-736cbb8c{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-736cbb96{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-736cbb97{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-736cbba0{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-736cbba1{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-7370df50'><thead><tr style="overflow-wrap:break-word;"><th class="cl-736cbb82"><p class="cl-736c9f58"><span class="cl-736a370e">mpg</span></p></th><th class="cl-736cbb82"><p class="cl-736c9f58"><span class="cl-736a370e">cyl</span></p></th><th class="cl-736cbb8c"><p class="cl-736c9f58"><span class="cl-736a370e">hp</span></p></th><th class="cl-736cbb96"><p class="cl-736c9f58"><span class="cl-736a370e">gear</span></p></th><th class="cl-736cbb96"><p class="cl-736c9f58"><span class="cl-736a370e">carb</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">21.0</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">6</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">110</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">21.0</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">6</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">110</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">22.8</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">93</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">21.4</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">6</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">110</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">3</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">18.7</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">8</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">175</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">3</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">18.1</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">6</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">105</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">3</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">14.3</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">8</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">245</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">3</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">24.4</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">62</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">22.8</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">95</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">19.2</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">6</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">123</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">17.8</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">6</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">123</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">16.4</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">8</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">180</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">3</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">17.3</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">8</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">180</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">3</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">15.2</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">8</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">180</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">3</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">10.4</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">8</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">205</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">3</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">10.4</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">8</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">215</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">3</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">14.7</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">8</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">230</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">3</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">32.4</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">66</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">30.4</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">52</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">33.9</span></p></td><td class="cl-736cbb97"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba0"><p class="cl-736c9f62"><span class="cl-736a370e">65</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">4</span></p></td><td class="cl-736cbba1"><p class="cl-736c9f62"><span class="cl-736a370e">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div>
#>   <button onclick="showPage(1)">1</button>
#>   <button onclick="showPage(2)">2</button>
#>   <button onclick="showPage(3)">3</button>
#> </div>
#> <script>
#>   function showPage(pageNum) {
#>     var pages = document.querySelectorAll('.page');
#>     pages.forEach(page => page.style.display = 'none');
#>     document.getElementById('page' + pageNum).style.display = 'block';
#>   }
#> </script>