Set column widths using percent

Extraction of flextable print method with special handling of clintable pages and

Usage

clin_col_widths(x, ...)

Arguments

x

A clintable object

...

Named parameters where the names are columns in the flextable and the values are decimals representing the percent of total width of the table

Value

A clintable object

Examples

ct <- clintable(mtcars)

ct <- clin_alt_pages(
  ct,
  key_cols = c("mpg", "cyl", "hp"),
  col_groups = list(
    c("disp", "drat", "wt"),
    c("qsec", "vs", "am"),
    c("gear", "carb")
  )
) |>
  clin_col_widths(mpg = .2, cyl = .2, disp = .15, vs = .15)

print(ct)
#> NOTE: Alternating pages were set, but no selection for row wise pagination was configured Defaulting to 20 rows per page.
#> <div id="page1" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-730208bc{}.cl-72fb0436{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-72fe394e{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-72fe3958{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-72fe54ec{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-72fe54f6{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-72fe54f7{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-72fe54f8{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-730208bc'><thead><tr style="overflow-wrap:break-word;"><th class="cl-72fe54ec"><p class="cl-72fe394e"><span class="cl-72fb0436">mpg</span></p></th><th class="cl-72fe54ec"><p class="cl-72fe394e"><span class="cl-72fb0436">cyl</span></p></th><th class="cl-72fe54f6"><p class="cl-72fe394e"><span class="cl-72fb0436">hp</span></p></th><th class="cl-72fe54f6"><p class="cl-72fe394e"><span class="cl-72fb0436">disp</span></p></th><th class="cl-72fe54f6"><p class="cl-72fe394e"><span class="cl-72fb0436">drat</span></p></th><th class="cl-72fe54f6"><p class="cl-72fe394e"><span class="cl-72fb0436">wt</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">21.0</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">6</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">110</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">160.0</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.90</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">2.620</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">21.0</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">6</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">110</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">160.0</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.90</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">2.875</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">22.8</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">4</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">93</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">108.0</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.85</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">2.320</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">21.4</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">6</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">110</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">258.0</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.08</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.215</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">18.7</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">8</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">175</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">360.0</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.15</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">18.1</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">6</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">105</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">225.0</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">2.76</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.460</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">14.3</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">8</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">245</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">360.0</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.21</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.570</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">24.4</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">4</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">62</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">146.7</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.69</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.190</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">22.8</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">4</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">95</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">140.8</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.92</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.150</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">19.2</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">6</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">123</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">167.6</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.92</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">17.8</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">6</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">123</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">167.6</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.92</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">16.4</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">8</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">180</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">275.8</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.07</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">4.070</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">17.3</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">8</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">180</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">275.8</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.07</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.730</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">15.2</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">8</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">180</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">275.8</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.07</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.780</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">10.4</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">8</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">205</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">472.0</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">2.93</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">5.250</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">10.4</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">8</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">215</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">460.0</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.00</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">5.424</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">14.7</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">8</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">230</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">440.0</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">3.23</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">5.345</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">32.4</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">4</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">66</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">78.7</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">4.08</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">2.200</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">30.4</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">4</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">52</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">75.7</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">4.93</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">1.615</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">33.9</span></p></td><td class="cl-72fe54f7"><p class="cl-72fe3958"><span class="cl-72fb0436">4</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">65</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">71.1</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">4.22</span></p></td><td class="cl-72fe54f8"><p class="cl-72fe3958"><span class="cl-72fb0436">1.835</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page2" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-730df96a{}.cl-7306def0{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-73095234{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-7309523e{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-73096cc4{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73096cce{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73096ccf{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73096cd0{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73096cd8{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73096cd9{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-730df96a'><thead><tr style="overflow-wrap:break-word;"><th class="cl-73096cc4"><p class="cl-73095234"><span class="cl-7306def0">mpg</span></p></th><th class="cl-73096cc4"><p class="cl-73095234"><span class="cl-7306def0">cyl</span></p></th><th class="cl-73096cce"><p class="cl-73095234"><span class="cl-7306def0">hp</span></p></th><th class="cl-73096ccf"><p class="cl-73095234"><span class="cl-7306def0">qsec</span></p></th><th class="cl-73096cce"><p class="cl-73095234"><span class="cl-7306def0">vs</span></p></th><th class="cl-73096ccf"><p class="cl-73095234"><span class="cl-7306def0">am</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">21.0</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">6</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">110</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">16.46</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">21.0</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">6</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">110</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">17.02</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">22.8</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">4</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">93</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">18.61</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">21.4</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">6</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">110</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">19.44</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">18.7</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">8</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">175</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">17.02</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">18.1</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">6</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">105</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">20.22</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">14.3</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">8</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">245</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">15.84</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">24.4</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">4</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">62</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">20.00</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">22.8</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">4</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">95</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">22.90</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">19.2</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">6</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">123</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">18.30</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">17.8</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">6</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">123</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">18.90</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">16.4</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">8</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">180</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">17.40</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">17.3</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">8</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">180</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">17.60</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">15.2</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">8</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">180</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">18.00</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">10.4</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">8</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">205</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">17.98</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">10.4</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">8</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">215</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">17.82</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">14.7</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">8</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">230</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">17.42</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">32.4</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">4</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">66</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">19.47</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">30.4</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">4</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">52</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">18.52</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">33.9</span></p></td><td class="cl-73096cd0"><p class="cl-7309523e"><span class="cl-7306def0">4</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">65</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">19.90</span></p></td><td class="cl-73096cd8"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td><td class="cl-73096cd9"><p class="cl-7309523e"><span class="cl-7306def0">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page3" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-7318e924{}.cl-73129146{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-731551ec{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-731551f6{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-73156da8{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73156db2{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73156dbc{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73156dbd{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73156dbe{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-73156dc6{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-7318e924'><thead><tr style="overflow-wrap:break-word;"><th class="cl-73156da8"><p class="cl-731551ec"><span class="cl-73129146">mpg</span></p></th><th class="cl-73156da8"><p class="cl-731551ec"><span class="cl-73129146">cyl</span></p></th><th class="cl-73156db2"><p class="cl-731551ec"><span class="cl-73129146">hp</span></p></th><th class="cl-73156dbc"><p class="cl-731551ec"><span class="cl-73129146">gear</span></p></th><th class="cl-73156dbc"><p class="cl-731551ec"><span class="cl-73129146">carb</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">21.0</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">6</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">110</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">21.0</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">6</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">110</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">22.8</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">93</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">21.4</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">6</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">110</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">3</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">18.7</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">8</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">175</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">3</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">18.1</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">6</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">105</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">3</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">14.3</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">8</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">245</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">3</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">24.4</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">62</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">22.8</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">95</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">19.2</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">6</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">123</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">17.8</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">6</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">123</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">16.4</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">8</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">180</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">3</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">17.3</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">8</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">180</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">3</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">15.2</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">8</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">180</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">3</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">10.4</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">8</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">205</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">3</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">10.4</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">8</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">215</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">3</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">14.7</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">8</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">230</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">3</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">32.4</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">66</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">30.4</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">52</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">33.9</span></p></td><td class="cl-73156dbd"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dbe"><p class="cl-731551f6"><span class="cl-73129146">65</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">4</span></p></td><td class="cl-73156dc6"><p class="cl-731551f6"><span class="cl-73129146">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div>
#>   <button onclick="showPage(1)">1</button>
#>   <button onclick="showPage(2)">2</button>
#>   <button onclick="showPage(3)">3</button>
#> </div>
#> <script>
#>   function showPage(pageNum) {
#>     var pages = document.querySelectorAll('.page');
#>     pages.forEach(page => page.style.display = 'none');
#>     document.getElementById('page' + pageNum).style.display = 'block';
#>   }
#> </script>