Set column widths using percent

Extraction of flextable print method with special handling of clintable pages and

Usage

clin_col_widths(x, ...)

Arguments

x

A clintable object

...

Named parameters where the names are columns in the flextable and the values are decimals representing the percent of total width of the table

Value

A clintable object

Examples

ct <- clintable(mtcars)

ct <- clin_alt_pages(
  ct,
  key_cols = c('mpg', 'cyl', 'hp'),
  col_groups = list(
    c('disp', 'drat', 'wt'),
    c('qsec', 'vs', 'am'),
    c('gear', 'carb')
  )
) |> 
clin_col_widths(mpg = .2, cyl=.2, disp=.15, vs=.15)

print(ct)
#> NOTE: Alternating pages were set, but no selection for row wise pagination was configured Defaulting to 20 rows per page.
#> <div id="page1" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-c6832444{}.cl-c67cb28a{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-c67f3eba{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-c67f3ec4{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-c67f5b8e{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-c67f5b98{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-c67f5ba2{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-c67f5ba3{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-c6832444'><thead><tr style="overflow-wrap:break-word;"><th class="cl-c67f5b8e"><p class="cl-c67f3eba"><span class="cl-c67cb28a">mpg</span></p></th><th class="cl-c67f5b8e"><p class="cl-c67f3eba"><span class="cl-c67cb28a">cyl</span></p></th><th class="cl-c67f5b98"><p class="cl-c67f3eba"><span class="cl-c67cb28a">hp</span></p></th><th class="cl-c67f5b98"><p class="cl-c67f3eba"><span class="cl-c67cb28a">disp</span></p></th><th class="cl-c67f5b98"><p class="cl-c67f3eba"><span class="cl-c67cb28a">drat</span></p></th><th class="cl-c67f5b98"><p class="cl-c67f3eba"><span class="cl-c67cb28a">wt</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">21.0</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">6</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">110</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">160.0</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.90</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">2.620</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">21.0</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">6</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">110</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">160.0</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.90</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">2.875</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">22.8</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">4</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">93</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">108.0</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.85</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">2.320</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">21.4</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">6</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">110</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">258.0</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.08</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.215</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">18.7</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">8</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">175</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">360.0</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.15</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">18.1</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">6</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">105</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">225.0</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">2.76</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.460</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">14.3</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">8</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">245</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">360.0</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.21</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.570</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">24.4</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">4</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">62</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">146.7</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.69</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.190</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">22.8</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">4</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">95</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">140.8</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.92</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.150</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">19.2</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">6</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">123</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">167.6</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.92</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">17.8</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">6</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">123</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">167.6</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.92</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">16.4</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">8</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">180</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">275.8</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.07</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">4.070</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">17.3</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">8</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">180</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">275.8</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.07</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.730</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">15.2</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">8</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">180</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">275.8</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.07</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.780</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">10.4</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">8</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">205</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">472.0</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">2.93</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">5.250</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">10.4</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">8</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">215</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">460.0</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.00</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">5.424</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">14.7</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">8</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">230</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">440.0</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">3.23</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">5.345</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">32.4</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">4</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">66</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">78.7</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">4.08</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">2.200</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">30.4</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">4</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">52</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">75.7</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">4.93</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">1.615</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">33.9</span></p></td><td class="cl-c67f5ba2"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">4</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">65</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">71.1</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">4.22</span></p></td><td class="cl-c67f5ba3"><p class="cl-c67f3ec4"><span class="cl-c67cb28a">1.835</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page2" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-c68ea760{}.cl-c6887908{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-c68ae6b6{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-c68ae6c0{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-c68b018c{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-c68b0196{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-c68b0197{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-c68b0198{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-c68b01a0{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-c68b01a1{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-c68ea760'><thead><tr style="overflow-wrap:break-word;"><th class="cl-c68b018c"><p class="cl-c68ae6b6"><span class="cl-c6887908">mpg</span></p></th><th class="cl-c68b018c"><p class="cl-c68ae6b6"><span class="cl-c6887908">cyl</span></p></th><th class="cl-c68b0196"><p class="cl-c68ae6b6"><span class="cl-c6887908">hp</span></p></th><th class="cl-c68b0197"><p class="cl-c68ae6b6"><span class="cl-c6887908">qsec</span></p></th><th class="cl-c68b0196"><p class="cl-c68ae6b6"><span class="cl-c6887908">vs</span></p></th><th class="cl-c68b0197"><p class="cl-c68ae6b6"><span class="cl-c6887908">am</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">21.0</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">6</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">110</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">16.46</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">21.0</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">6</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">110</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">17.02</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">22.8</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">4</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">93</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">18.61</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">21.4</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">6</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">110</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">19.44</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">18.7</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">8</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">175</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">17.02</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">18.1</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">6</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">105</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">20.22</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">14.3</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">8</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">245</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">15.84</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">24.4</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">4</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">62</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">20.00</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">22.8</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">4</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">95</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">22.90</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">19.2</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">6</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">123</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">18.30</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">17.8</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">6</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">123</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">18.90</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">16.4</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">8</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">180</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">17.40</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">17.3</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">8</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">180</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">17.60</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">15.2</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">8</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">180</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">18.00</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">10.4</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">8</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">205</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">17.98</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">10.4</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">8</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">215</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">17.82</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">14.7</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">8</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">230</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">17.42</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">32.4</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">4</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">66</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">19.47</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">30.4</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">4</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">52</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">18.52</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">33.9</span></p></td><td class="cl-c68b0198"><p class="cl-c68ae6c0"><span class="cl-c6887908">4</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">65</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">19.90</span></p></td><td class="cl-c68b01a0"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td><td class="cl-c68b01a1"><p class="cl-c68ae6c0"><span class="cl-c6887908">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page3" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-c699aeb2{}.cl-c6930f8a{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-c696175c{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-c6961770{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-c69631d8{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-c69631e2{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-c69631e3{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-c69631e4{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-c69631ec{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-c69631ed{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-c699aeb2'><thead><tr style="overflow-wrap:break-word;"><th class="cl-c69631d8"><p class="cl-c696175c"><span class="cl-c6930f8a">mpg</span></p></th><th class="cl-c69631d8"><p class="cl-c696175c"><span class="cl-c6930f8a">cyl</span></p></th><th class="cl-c69631e2"><p class="cl-c696175c"><span class="cl-c6930f8a">hp</span></p></th><th class="cl-c69631e3"><p class="cl-c696175c"><span class="cl-c6930f8a">gear</span></p></th><th class="cl-c69631e3"><p class="cl-c696175c"><span class="cl-c6930f8a">carb</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">21.0</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">6</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">110</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">21.0</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">6</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">110</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">22.8</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">93</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">21.4</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">6</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">110</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">3</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">18.7</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">8</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">175</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">3</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">18.1</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">6</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">105</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">3</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">14.3</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">8</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">245</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">3</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">24.4</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">62</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">22.8</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">95</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">19.2</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">6</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">123</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">17.8</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">6</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">123</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">16.4</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">8</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">180</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">3</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">17.3</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">8</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">180</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">3</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">15.2</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">8</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">180</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">3</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">10.4</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">8</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">205</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">3</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">10.4</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">8</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">215</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">3</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">14.7</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">8</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">230</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">3</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">32.4</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">66</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">30.4</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">52</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">33.9</span></p></td><td class="cl-c69631e4"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ec"><p class="cl-c6961770"><span class="cl-c6930f8a">65</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">4</span></p></td><td class="cl-c69631ed"><p class="cl-c6961770"><span class="cl-c6930f8a">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div>
#>   <button onclick="showPage(1)">1</button>
#>   <button onclick="showPage(2)">2</button>
#>   <button onclick="showPage(3)">3</button>
#> </div>
#> <script>
#>   function showPage(pageNum) {
#>     var pages = document.querySelectorAll('.page');
#>     pages.forEach(page => page.style.display = 'none');
#>     document.getElementById('page' + pageNum).style.display = 'block';
#>   }
#> </script>