Set column widths using percent

Extraction of flextable print method with special handling of clintable pages and

Usage

clin_col_widths(x, ...)

Arguments

x

A clintable object

...

Named parameters where the names are columns in the flextable and the values are decimals representing the percent of total width of the table

Value

A clintable object

Examples

ct <- clintable(mtcars)

ct <- clin_alt_pages(
  ct,
  key_cols = c("mpg", "cyl", "hp"),
  col_groups = list(
    c("disp", "drat", "wt"),
    c("qsec", "vs", "am"),
    c("gear", "carb")
  )
) |>
  clin_col_widths(mpg = .2, cyl = .2, disp = .15, vs = .15)

print(ct)
#> NOTE: Alternating pages were set, but no selection for row wise pagination was configured Defaulting to 20 rows per page.
#> <div id="page1" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-b05b3c22{}.cl-b0540682{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-b0575594{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-b057559e{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-b0577204{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-b057720e{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-b0577218{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-b0577219{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-b05b3c22'><thead><tr style="overflow-wrap:break-word;"><th class="cl-b0577204"><p class="cl-b0575594"><span class="cl-b0540682">mpg</span></p></th><th class="cl-b0577204"><p class="cl-b0575594"><span class="cl-b0540682">cyl</span></p></th><th class="cl-b057720e"><p class="cl-b0575594"><span class="cl-b0540682">hp</span></p></th><th class="cl-b057720e"><p class="cl-b0575594"><span class="cl-b0540682">disp</span></p></th><th class="cl-b057720e"><p class="cl-b0575594"><span class="cl-b0540682">drat</span></p></th><th class="cl-b057720e"><p class="cl-b0575594"><span class="cl-b0540682">wt</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">21.0</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">6</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">110</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">160.0</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.90</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">2.620</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">21.0</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">6</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">110</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">160.0</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.90</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">2.875</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">22.8</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">4</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">93</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">108.0</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.85</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">2.320</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">21.4</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">6</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">110</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">258.0</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.08</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.215</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">18.7</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">8</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">175</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">360.0</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.15</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">18.1</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">6</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">105</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">225.0</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">2.76</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.460</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">14.3</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">8</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">245</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">360.0</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.21</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.570</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">24.4</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">4</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">62</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">146.7</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.69</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.190</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">22.8</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">4</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">95</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">140.8</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.92</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.150</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">19.2</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">6</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">123</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">167.6</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.92</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">17.8</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">6</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">123</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">167.6</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.92</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">16.4</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">8</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">180</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">275.8</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.07</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">4.070</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">17.3</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">8</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">180</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">275.8</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.07</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.730</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">15.2</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">8</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">180</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">275.8</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.07</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.780</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">10.4</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">8</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">205</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">472.0</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">2.93</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">5.250</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">10.4</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">8</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">215</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">460.0</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.00</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">5.424</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">14.7</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">8</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">230</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">440.0</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">3.23</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">5.345</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">32.4</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">4</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">66</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">78.7</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">4.08</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">2.200</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">30.4</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">4</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">52</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">75.7</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">4.93</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">1.615</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">33.9</span></p></td><td class="cl-b0577218"><p class="cl-b057559e"><span class="cl-b0540682">4</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">65</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">71.1</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">4.22</span></p></td><td class="cl-b0577219"><p class="cl-b057559e"><span class="cl-b0540682">1.835</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page2" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-b066f27e{}.cl-b05fd32c{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-b0625c1e{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-b0625c28{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-b0627730{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-b062773a{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-b062773b{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-b0627744{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-b0627745{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-b0627746{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-b066f27e'><thead><tr style="overflow-wrap:break-word;"><th class="cl-b0627730"><p class="cl-b0625c1e"><span class="cl-b05fd32c">mpg</span></p></th><th class="cl-b0627730"><p class="cl-b0625c1e"><span class="cl-b05fd32c">cyl</span></p></th><th class="cl-b062773a"><p class="cl-b0625c1e"><span class="cl-b05fd32c">hp</span></p></th><th class="cl-b062773b"><p class="cl-b0625c1e"><span class="cl-b05fd32c">qsec</span></p></th><th class="cl-b062773a"><p class="cl-b0625c1e"><span class="cl-b05fd32c">vs</span></p></th><th class="cl-b062773b"><p class="cl-b0625c1e"><span class="cl-b05fd32c">am</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">21.0</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">6</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">110</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">16.46</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">21.0</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">6</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">110</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">17.02</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">22.8</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">4</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">93</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">18.61</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">21.4</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">6</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">110</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">19.44</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">18.7</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">8</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">175</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">17.02</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">18.1</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">6</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">105</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">20.22</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">14.3</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">8</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">245</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">15.84</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">24.4</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">4</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">62</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">20.00</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">22.8</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">4</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">95</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">22.90</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">19.2</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">6</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">123</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">18.30</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">17.8</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">6</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">123</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">18.90</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">16.4</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">8</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">180</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">17.40</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">17.3</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">8</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">180</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">17.60</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">15.2</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">8</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">180</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">18.00</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">10.4</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">8</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">205</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">17.98</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">10.4</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">8</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">215</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">17.82</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">14.7</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">8</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">230</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">17.42</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">32.4</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">4</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">66</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">19.47</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">30.4</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">4</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">52</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">18.52</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">33.9</span></p></td><td class="cl-b0627744"><p class="cl-b0625c28"><span class="cl-b05fd32c">4</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">65</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">19.90</span></p></td><td class="cl-b0627745"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td><td class="cl-b0627746"><p class="cl-b0625c28"><span class="cl-b05fd32c">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page3" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-b07234c2{}.cl-b06b47a2{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-b06dc9fa{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-b06dca04{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-b06de53e{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-b06de53f{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-b06de548{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-b06de549{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-b06de54a{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-b06de552{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-b07234c2'><thead><tr style="overflow-wrap:break-word;"><th class="cl-b06de53e"><p class="cl-b06dc9fa"><span class="cl-b06b47a2">mpg</span></p></th><th class="cl-b06de53e"><p class="cl-b06dc9fa"><span class="cl-b06b47a2">cyl</span></p></th><th class="cl-b06de53f"><p class="cl-b06dc9fa"><span class="cl-b06b47a2">hp</span></p></th><th class="cl-b06de548"><p class="cl-b06dc9fa"><span class="cl-b06b47a2">gear</span></p></th><th class="cl-b06de548"><p class="cl-b06dc9fa"><span class="cl-b06b47a2">carb</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">21.0</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">6</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">110</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">21.0</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">6</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">110</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">22.8</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">93</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">21.4</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">6</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">110</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">3</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">18.7</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">8</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">175</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">3</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">18.1</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">6</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">105</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">3</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">14.3</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">8</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">245</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">3</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">24.4</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">62</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">22.8</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">95</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">19.2</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">6</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">123</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">17.8</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">6</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">123</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">16.4</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">8</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">180</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">3</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">17.3</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">8</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">180</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">3</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">15.2</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">8</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">180</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">3</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">10.4</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">8</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">205</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">3</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">10.4</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">8</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">215</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">3</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">14.7</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">8</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">230</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">3</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">32.4</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">66</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">30.4</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">52</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">33.9</span></p></td><td class="cl-b06de549"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de54a"><p class="cl-b06dca04"><span class="cl-b06b47a2">65</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">4</span></p></td><td class="cl-b06de552"><p class="cl-b06dca04"><span class="cl-b06b47a2">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div>
#>   <button onclick="showPage(1)">1</button>
#>   <button onclick="showPage(2)">2</button>
#>   <button onclick="showPage(3)">3</button>
#> </div>
#> <script>
#>   function showPage(pageNum) {
#>     var pages = document.querySelectorAll('.page');
#>     pages.forEach(page => page.style.display = 'none');
#>     document.getElementById('page' + pageNum).style.display = 'block';
#>   }
#> </script>