Set column widths using percent

Extraction of flextable print method with special handling of clintable pages and

Usage

clin_col_widths(x, ...)

Arguments

x

A clintable object

...

Named parameters where the names are columns in the flextable and the values are decimals representing the percent of total width of the table

Value

A clintable object

Examples

ct <- clintable(mtcars)

ct <- clin_alt_pages(
  ct,
  key_cols = c("mpg", "cyl", "hp"),
  col_groups = list(
    c("disp", "drat", "wt"),
    c("qsec", "vs", "am"),
    c("gear", "carb")
  )
) |>
  clin_col_widths(mpg = .2, cyl = .2, disp = .15, vs = .15)

print(ct)
#> NOTE: Alternating pages were set, but no selection for row wise pagination was configured Defaulting to 20 rows per page.
#> <div id="page1" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-57b9503a{}.cl-57b169ce{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-57b51d12{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-57b51d26{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-57b53a72{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-57b53a73{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-57b53a7c{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-57b53a86{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-57b9503a'><thead><tr style="overflow-wrap:break-word;"><th class="cl-57b53a72"><p class="cl-57b51d12"><span class="cl-57b169ce">mpg</span></p></th><th class="cl-57b53a72"><p class="cl-57b51d12"><span class="cl-57b169ce">cyl</span></p></th><th class="cl-57b53a73"><p class="cl-57b51d12"><span class="cl-57b169ce">hp</span></p></th><th class="cl-57b53a73"><p class="cl-57b51d12"><span class="cl-57b169ce">disp</span></p></th><th class="cl-57b53a73"><p class="cl-57b51d12"><span class="cl-57b169ce">drat</span></p></th><th class="cl-57b53a73"><p class="cl-57b51d12"><span class="cl-57b169ce">wt</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">21.0</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">6</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">110</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">160.0</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.90</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">2.620</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">21.0</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">6</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">110</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">160.0</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.90</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">2.875</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">22.8</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">4</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">93</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">108.0</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.85</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">2.320</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">21.4</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">6</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">110</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">258.0</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.08</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.215</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">18.7</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">8</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">175</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">360.0</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.15</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">18.1</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">6</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">105</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">225.0</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">2.76</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.460</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">14.3</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">8</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">245</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">360.0</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.21</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.570</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">24.4</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">4</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">62</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">146.7</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.69</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.190</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">22.8</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">4</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">95</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">140.8</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.92</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.150</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">19.2</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">6</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">123</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">167.6</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.92</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">17.8</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">6</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">123</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">167.6</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.92</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.440</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">16.4</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">8</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">180</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">275.8</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.07</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">4.070</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">17.3</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">8</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">180</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">275.8</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.07</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.730</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">15.2</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">8</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">180</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">275.8</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.07</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.780</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">10.4</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">8</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">205</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">472.0</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">2.93</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">5.250</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">10.4</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">8</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">215</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">460.0</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.00</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">5.424</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">14.7</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">8</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">230</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">440.0</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">3.23</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">5.345</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">32.4</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">4</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">66</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">78.7</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">4.08</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">2.200</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">30.4</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">4</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">52</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">75.7</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">4.93</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">1.615</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">33.9</span></p></td><td class="cl-57b53a7c"><p class="cl-57b51d26"><span class="cl-57b169ce">4</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">65</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">71.1</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">4.22</span></p></td><td class="cl-57b53a86"><p class="cl-57b51d26"><span class="cl-57b169ce">1.835</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page2" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-57c6b54a{}.cl-57be87b2{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-57c16cca{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-57c16cde{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-57c18bf6{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-57c18c00{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-57c18c01{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-57c18c0a{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-57c18c0b{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-57c18c0c{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-57c6b54a'><thead><tr style="overflow-wrap:break-word;"><th class="cl-57c18bf6"><p class="cl-57c16cca"><span class="cl-57be87b2">mpg</span></p></th><th class="cl-57c18bf6"><p class="cl-57c16cca"><span class="cl-57be87b2">cyl</span></p></th><th class="cl-57c18c00"><p class="cl-57c16cca"><span class="cl-57be87b2">hp</span></p></th><th class="cl-57c18c01"><p class="cl-57c16cca"><span class="cl-57be87b2">qsec</span></p></th><th class="cl-57c18c00"><p class="cl-57c16cca"><span class="cl-57be87b2">vs</span></p></th><th class="cl-57c18c01"><p class="cl-57c16cca"><span class="cl-57be87b2">am</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">21.0</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">6</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">110</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">16.46</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">21.0</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">6</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">110</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">17.02</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">22.8</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">4</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">93</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">18.61</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">21.4</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">6</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">110</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">19.44</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">18.7</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">8</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">175</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">17.02</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">18.1</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">6</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">105</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">20.22</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">14.3</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">8</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">245</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">15.84</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">24.4</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">4</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">62</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">20.00</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">22.8</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">4</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">95</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">22.90</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">19.2</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">6</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">123</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">18.30</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">17.8</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">6</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">123</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">18.90</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">16.4</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">8</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">180</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">17.40</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">17.3</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">8</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">180</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">17.60</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">15.2</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">8</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">180</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">18.00</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">10.4</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">8</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">205</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">17.98</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">10.4</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">8</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">215</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">17.82</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">14.7</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">8</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">230</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">17.42</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">0</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">32.4</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">4</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">66</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">19.47</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">30.4</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">4</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">52</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">18.52</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">33.9</span></p></td><td class="cl-57c18c0a"><p class="cl-57c16cde"><span class="cl-57be87b2">4</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">65</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">19.90</span></p></td><td class="cl-57c18c0b"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td><td class="cl-57c18c0c"><p class="cl-57c16cde"><span class="cl-57be87b2">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div id="page3" style="display:none;" class="page">
#>   <style></style>
#>   <div class="tabwid"><style>.cl-57d23550{}.cl-57cb64a0{font-family:'Courier New';font-size:9pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;}.cl-57ce1448{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:9pt;padding-top:9pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-57ce1452{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:0.1pt;padding-top:0.1pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;}.cl-57ce30cc{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-57ce30d6{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-57ce30e0{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 1pt solid rgba(102, 102, 102, 1.00);border-top: 1pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-57ce30e1{width:1.945in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-57ce30ea{width:1.459in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}.cl-57ce30eb{width:2.188in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;}</style><table data-quarto-disable-processing='true' class='cl-57d23550'><thead><tr style="overflow-wrap:break-word;"><th class="cl-57ce30cc"><p class="cl-57ce1448"><span class="cl-57cb64a0">mpg</span></p></th><th class="cl-57ce30cc"><p class="cl-57ce1448"><span class="cl-57cb64a0">cyl</span></p></th><th class="cl-57ce30d6"><p class="cl-57ce1448"><span class="cl-57cb64a0">hp</span></p></th><th class="cl-57ce30e0"><p class="cl-57ce1448"><span class="cl-57cb64a0">gear</span></p></th><th class="cl-57ce30e0"><p class="cl-57ce1448"><span class="cl-57cb64a0">carb</span></p></th></tr></thead><tbody><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">21.0</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">6</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">110</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">21.0</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">6</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">110</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">22.8</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">93</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">21.4</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">6</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">110</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">3</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">18.7</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">8</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">175</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">3</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">18.1</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">6</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">105</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">3</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">14.3</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">8</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">245</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">3</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">24.4</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">62</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">22.8</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">95</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">19.2</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">6</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">123</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">17.8</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">6</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">123</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">16.4</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">8</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">180</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">3</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">17.3</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">8</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">180</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">3</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">15.2</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">8</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">180</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">3</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">3</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">10.4</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">8</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">205</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">3</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">10.4</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">8</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">215</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">3</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">14.7</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">8</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">230</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">3</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">32.4</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">66</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">1</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">30.4</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">52</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">2</span></p></td></tr><tr style="overflow-wrap:break-word;"><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">33.9</span></p></td><td class="cl-57ce30e1"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30ea"><p class="cl-57ce1452"><span class="cl-57cb64a0">65</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">4</span></p></td><td class="cl-57ce30eb"><p class="cl-57ce1452"><span class="cl-57cb64a0">1</span></p></td></tr></tbody></table></div>
#> </div>
#> <div>
#>   <button onclick="showPage(1)">1</button>
#>   <button onclick="showPage(2)">2</button>
#>   <button onclick="showPage(3)">3</button>
#> </div>
#> <script>
#>   function showPage(pageNum) {
#>     var pages = document.querySelectorAll('.page');
#>     pages.forEach(page => page.style.display = 'none');
#>     document.getElementById('page' + pageNum).style.display = 'block';
#>   }
#> </script>