Configure a clintable to table by a grouping variable, which will be used as a label

Configure a clintable to table by a grouping variable, which will be used as a label

Usage

clin_group_by(x, group_by)

Arguments

x

A clintable object

group_by

A character vector of variable names which will be used for grouping and attached as a label above the table headers

Value

A clintable object

Examples

clintable(iris) |>
  clin_group_by('Species')

  

setosa

Sepal.Length

Sepal.Width

Petal.Length

Petal.Width

5.1

3.5

1.4

0.2

  

setosa

Sepal.Length

Sepal.Width

Petal.Length

Petal.Width

4.9

3.0

1.4

0.2

  

setosa

Sepal.Length

Sepal.Width

Petal.Length

Petal.Width

4.7

3.2

1.3

0.2