Create a clindoc object

These functions handle the conversion of a clintable object into a clindoc object.

Usage

clindoc(...)

as_clindoc(x)

Arguments

...

clintable objects to be converted. Or separately, a list of clintable objects

x

A clintable object to be converted.

Value

a clindoc object, inherited from an officer::rdocx object

Details

as_clindoc() is intended for a single clintable object and extracts all necessary title and footnote information to apply to the document

clindoc() can accept clintable objects as separate parameters or as a list of clintable objects. If a single clintable is passed, this function simply calls as_clindoc() internally.

When a multiple clintable objects are passed to clindoc(), titles and footnotes should be applied directly to the clindoc object using clin_add_title(), clin_add_footnote(), or clin_add_footnote_page(). Title and footnote information on the individual clintable objects will be ignored.

Examples

ct <- clintable(mtcars)

clindoc(ct)
#> rdocx document with 2 element(s)
#> 
#> * styles:
#>                 Normal              heading 1              heading 2 
#>            "paragraph"            "paragraph"            "paragraph" 
#>              heading 3 Default Paragraph Font           Normal Table 
#>            "paragraph"            "character"                "table" 
#>                No List                 strong               centered 
#>            "numbering"            "character"            "paragraph" 
#>         table_template    Light List Accent 2            Titre 1 Car 
#>                "table"                "table"            "character" 
#>            Titre 2 Car            Titre 3 Car          Image Caption 
#>            "character"            "character"            "paragraph" 
#>          Table Caption     Table Professional                  toc 1 
#>            "paragraph"                "table"            "paragraph" 
#>                  toc 2           Balloon Text    Texte de bulles Car 
#>            "paragraph"            "paragraph"            "character" 
#>           reference_id          graphic title            table title 
#>            "character"            "paragraph"            "paragraph" 
#> 
#> * Content at cursor location:
#>        row_id is_header cell_id  text col_span row_span content_type
#> 1.1         1      TRUE       1   mpg        1        1   table cell
#> 1.12        2     FALSE       1  21.0        1        1   table cell
#> 1.23        3     FALSE       1  21.0        1        1   table cell
#> 1.34        4     FALSE       1  22.8        1        1   table cell
#> 1.45        5     FALSE       1  21.4        1        1   table cell
#> 1.56        6     FALSE       1  18.7        1        1   table cell
#> 1.67        7     FALSE       1  18.1        1        1   table cell
#> 1.78        8     FALSE       1  14.3        1        1   table cell
#> 1.89        9     FALSE       1  24.4        1        1   table cell
#> 1.100      10     FALSE       1  22.8        1        1   table cell
#> 1.111      11     FALSE       1  19.2        1        1   table cell
#> 1.122      12     FALSE       1  17.8        1        1   table cell
#> 1.133      13     FALSE       1  16.4        1        1   table cell
#> 1.144      14     FALSE       1  17.3        1        1   table cell
#> 1.155      15     FALSE       1  15.2        1        1   table cell
#> 1.166      16     FALSE       1  10.4        1        1   table cell
#> 1.177      17     FALSE       1  10.4        1        1   table cell
#> 1.188      18     FALSE       1  14.7        1        1   table cell
#> 1.199      19     FALSE       1  32.4        1        1   table cell
#> 1.210      20     FALSE       1  30.4        1        1   table cell
#> 1.221      21     FALSE       1  33.9        1        1   table cell
#> 1.232      22     FALSE       1  21.5        1        1   table cell
#> 1.243      23     FALSE       1  15.5        1        1   table cell
#> 1.254      24     FALSE       1  15.2        1        1   table cell
#> 1.265      25     FALSE       1  13.3        1        1   table cell
#> 1.276      26     FALSE       1  19.2        1        1   table cell
#> 1.287      27     FALSE       1  27.3        1        1   table cell
#> 1.298      28     FALSE       1  26.0        1        1   table cell
#> 1.309      29     FALSE       1  30.4        1        1   table cell
#> 1.320      30     FALSE       1  15.8        1        1   table cell
#> 1.331      31     FALSE       1  19.7        1        1   table cell
#> 1.342      32     FALSE       1  15.0        1        1   table cell
#> 1.353      33     FALSE       1  21.4        1        1   table cell
#> 2.2         1      TRUE       2   cyl        1        1   table cell
#> 2.13        2     FALSE       2     6        1        1   table cell
#> 2.24        3     FALSE       2     6        1        1   table cell
#> 2.35        4     FALSE       2     4        1        1   table cell
#> 2.46        5     FALSE       2     6        1        1   table cell
#> 2.57        6     FALSE       2     8        1        1   table cell
#> 2.68        7     FALSE       2     6        1        1   table cell
#> 2.79        8     FALSE       2     8        1        1   table cell
#> 2.90        9     FALSE       2     4        1        1   table cell
#> 2.101      10     FALSE       2     4        1        1   table cell
#> 2.112      11     FALSE       2     6        1        1   table cell
#> 2.123      12     FALSE       2     6        1        1   table cell
#> 2.134      13     FALSE       2     8        1        1   table cell
#> 2.145      14     FALSE       2     8        1        1   table cell
#> 2.156      15     FALSE       2     8        1        1   table cell
#> 2.167      16     FALSE       2     8        1        1   table cell
#> 2.178      17     FALSE       2     8        1        1   table cell
#> 2.189      18     FALSE       2     8        1        1   table cell
#> 2.200      19     FALSE       2     4        1        1   table cell
#> 2.211      20     FALSE       2     4        1        1   table cell
#> 2.222      21     FALSE       2     4        1        1   table cell
#> 2.233      22     FALSE       2     4        1        1   table cell
#> 2.244      23     FALSE       2     8        1        1   table cell
#> 2.255      24     FALSE       2     8        1        1   table cell
#> 2.266      25     FALSE       2     8        1        1   table cell
#> 2.277      26     FALSE       2     8        1        1   table cell
#> 2.288      27     FALSE       2     4        1        1   table cell
#> 2.299      28     FALSE       2     4        1        1   table cell
#> 2.310      29     FALSE       2     4        1        1   table cell
#> 2.321      30     FALSE       2     8        1        1   table cell
#> 2.332      31     FALSE       2     6        1        1   table cell
#> 2.343      32     FALSE       2     8        1        1   table cell
#> 2.354      33     FALSE       2     4        1        1   table cell
#> 3.3         1      TRUE       3  disp        1        1   table cell
#> 3.14        2     FALSE       3 160.0        1        1   table cell
#> 3.25        3     FALSE       3 160.0        1        1   table cell
#> 3.36        4     FALSE       3 108.0        1        1   table cell
#> 3.47        5     FALSE       3 258.0        1        1   table cell
#> 3.58        6     FALSE       3 360.0        1        1   table cell
#> 3.69        7     FALSE       3 225.0        1        1   table cell
#> 3.80        8     FALSE       3 360.0        1        1   table cell
#> 3.91        9     FALSE       3 146.7        1        1   table cell
#> 3.102      10     FALSE       3 140.8        1        1   table cell
#> 3.113      11     FALSE       3 167.6        1        1   table cell
#> 3.124      12     FALSE       3 167.6        1        1   table cell
#> 3.135      13     FALSE       3 275.8        1        1   table cell
#> 3.146      14     FALSE       3 275.8        1        1   table cell
#> 3.157      15     FALSE       3 275.8        1        1   table cell
#> 3.168      16     FALSE       3 472.0        1        1   table cell
#> 3.179      17     FALSE       3 460.0        1        1   table cell
#> 3.190      18     FALSE       3 440.0        1        1   table cell
#> 3.201      19     FALSE       3  78.7        1        1   table cell
#> 3.212      20     FALSE       3  75.7        1        1   table cell
#> 3.223      21     FALSE       3  71.1        1        1   table cell
#> 3.234      22     FALSE       3 120.1        1        1   table cell
#> 3.245      23     FALSE       3 318.0        1        1   table cell
#> 3.256      24     FALSE       3 304.0        1        1   table cell
#> 3.267      25     FALSE       3 350.0        1        1   table cell
#> 3.278      26     FALSE       3 400.0        1        1   table cell
#> 3.289      27     FALSE       3  79.0        1        1   table cell
#> 3.300      28     FALSE       3 120.3        1        1   table cell
#> 3.311      29     FALSE       3  95.1        1        1   table cell
#> 3.322      30     FALSE       3 351.0        1        1   table cell
#> 3.333      31     FALSE       3 145.0        1        1   table cell
#> 3.344      32     FALSE       3 301.0        1        1   table cell
#> 3.355      33     FALSE       3 121.0        1        1   table cell
#> 4.4         1      TRUE       4    hp        1        1   table cell
#> 4.15        2     FALSE       4   110        1        1   table cell
#> 4.26        3     FALSE       4   110        1        1   table cell
#> 4.37        4     FALSE       4    93        1        1   table cell
#> 4.48        5     FALSE       4   110        1        1   table cell
#> 4.59        6     FALSE       4   175        1        1   table cell
#> 4.70        7     FALSE       4   105        1        1   table cell
#> 4.81        8     FALSE       4   245        1        1   table cell
#> 4.92        9     FALSE       4    62        1        1   table cell
#> 4.103      10     FALSE       4    95        1        1   table cell
#> 4.114      11     FALSE       4   123        1        1   table cell
#> 4.125      12     FALSE       4   123        1        1   table cell
#> 4.136      13     FALSE       4   180        1        1   table cell
#> 4.147      14     FALSE       4   180        1        1   table cell
#> 4.158      15     FALSE       4   180        1        1   table cell
#> 4.169      16     FALSE       4   205        1        1   table cell
#> 4.180      17     FALSE       4   215        1        1   table cell
#> 4.191      18     FALSE       4   230        1        1   table cell
#> 4.202      19     FALSE       4    66        1        1   table cell
#> 4.213      20     FALSE       4    52        1        1   table cell
#> 4.224      21     FALSE       4    65        1        1   table cell
#> 4.235      22     FALSE       4    97        1        1   table cell
#> 4.246      23     FALSE       4   150        1        1   table cell
#> 4.257      24     FALSE       4   150        1        1   table cell
#> 4.268      25     FALSE       4   245        1        1   table cell
#> 4.279      26     FALSE       4   175        1        1   table cell
#> 4.290      27     FALSE       4    66        1        1   table cell
#> 4.301      28     FALSE       4    91        1        1   table cell
#> 4.312      29     FALSE       4   113        1        1   table cell
#> 4.323      30     FALSE       4   264        1        1   table cell
#> 4.334      31     FALSE       4   175        1        1   table cell
#> 4.345      32     FALSE       4   335        1        1   table cell
#> 4.356      33     FALSE       4   109        1        1   table cell
#> 5.5         1      TRUE       5  drat        1        1   table cell
#> 5.16        2     FALSE       5  3.90        1        1   table cell
#> 5.27        3     FALSE       5  3.90        1        1   table cell
#> 5.38        4     FALSE       5  3.85        1        1   table cell
#> 5.49        5     FALSE       5  3.08        1        1   table cell
#> 5.60        6     FALSE       5  3.15        1        1   table cell
#> 5.71        7     FALSE       5  2.76        1        1   table cell
#> 5.82        8     FALSE       5  3.21        1        1   table cell
#> 5.93        9     FALSE       5  3.69        1        1   table cell
#> 5.104      10     FALSE       5  3.92        1        1   table cell
#> 5.115      11     FALSE       5  3.92        1        1   table cell
#> 5.126      12     FALSE       5  3.92        1        1   table cell
#> 5.137      13     FALSE       5  3.07        1        1   table cell
#> 5.148      14     FALSE       5  3.07        1        1   table cell
#> 5.159      15     FALSE       5  3.07        1        1   table cell
#> 5.170      16     FALSE       5  2.93        1        1   table cell
#> 5.181      17     FALSE       5  3.00        1        1   table cell
#> 5.192      18     FALSE       5  3.23        1        1   table cell
#> 5.203      19     FALSE       5  4.08        1        1   table cell
#> 5.214      20     FALSE       5  4.93        1        1   table cell
#> 5.225      21     FALSE       5  4.22        1        1   table cell
#> 5.236      22     FALSE       5  3.70        1        1   table cell
#> 5.247      23     FALSE       5  2.76        1        1   table cell
#> 5.258      24     FALSE       5  3.15        1        1   table cell
#> 5.269      25     FALSE       5  3.73        1        1   table cell
#> 5.280      26     FALSE       5  3.08        1        1   table cell
#> 5.291      27     FALSE       5  4.08        1        1   table cell
#> 5.302      28     FALSE       5  4.43        1        1   table cell
#> 5.313      29     FALSE       5  3.77        1        1   table cell
#> 5.324      30     FALSE       5  4.22        1        1   table cell
#> 5.335      31     FALSE       5  3.62        1        1   table cell
#> 5.346      32     FALSE       5  3.54        1        1   table cell
#> 5.357      33     FALSE       5  4.11        1        1   table cell
#> 6.6         1      TRUE       6    wt        1        1   table cell
#> 6.17        2     FALSE       6 2.620        1        1   table cell
#> 6.28        3     FALSE       6 2.875        1        1   table cell
#> 6.39        4     FALSE       6 2.320        1        1   table cell
#> 6.50        5     FALSE       6 3.215        1        1   table cell
#> 6.61        6     FALSE       6 3.440        1        1   table cell
#> 6.72        7     FALSE       6 3.460        1        1   table cell
#> 6.83        8     FALSE       6 3.570        1        1   table cell
#> 6.94        9     FALSE       6 3.190        1        1   table cell
#> 6.105      10     FALSE       6 3.150        1        1   table cell
#> 6.116      11     FALSE       6 3.440        1        1   table cell
#> 6.127      12     FALSE       6 3.440        1        1   table cell
#> 6.138      13     FALSE       6 4.070        1        1   table cell
#> 6.149      14     FALSE       6 3.730        1        1   table cell
#> 6.160      15     FALSE       6 3.780        1        1   table cell
#> 6.171      16     FALSE       6 5.250        1        1   table cell
#> 6.182      17     FALSE       6 5.424        1        1   table cell
#> 6.193      18     FALSE       6 5.345        1        1   table cell
#> 6.204      19     FALSE       6 2.200        1        1   table cell
#> 6.215      20     FALSE       6 1.615        1        1   table cell
#> 6.226      21     FALSE       6 1.835        1        1   table cell
#> 6.237      22     FALSE       6 2.465        1        1   table cell
#> 6.248      23     FALSE       6 3.520        1        1   table cell
#> 6.259      24     FALSE       6 3.435        1        1   table cell
#> 6.270      25     FALSE       6 3.840        1        1   table cell
#> 6.281      26     FALSE       6 3.845        1        1   table cell
#> 6.292      27     FALSE       6 1.935        1        1   table cell
#> 6.303      28     FALSE       6 2.140        1        1   table cell
#> 6.314      29     FALSE       6 1.513        1        1   table cell
#> 6.325      30     FALSE       6 3.170        1        1   table cell
#> 6.336      31     FALSE       6 2.770        1        1   table cell
#> 6.347      32     FALSE       6 3.570        1        1   table cell
#> 6.358      33     FALSE       6 2.780        1        1   table cell
#> 7.7         1      TRUE       7  qsec        1        1   table cell
#> 7.18        2     FALSE       7 16.46        1        1   table cell
#> 7.29        3     FALSE       7 17.02        1        1   table cell
#> 7.40        4     FALSE       7 18.61        1        1   table cell
#> 7.51        5     FALSE       7 19.44        1        1   table cell
#> 7.62        6     FALSE       7 17.02        1        1   table cell
#> 7.73        7     FALSE       7 20.22        1        1   table cell
#> 7.84        8     FALSE       7 15.84        1        1   table cell
#> 7.95        9     FALSE       7 20.00        1        1   table cell
#> 7.106      10     FALSE       7 22.90        1        1   table cell
#> 7.117      11     FALSE       7 18.30        1        1   table cell
#> 7.128      12     FALSE       7 18.90        1        1   table cell
#> 7.139      13     FALSE       7 17.40        1        1   table cell
#> 7.150      14     FALSE       7 17.60        1        1   table cell
#> 7.161      15     FALSE       7 18.00        1        1   table cell
#> 7.172      16     FALSE       7 17.98        1        1   table cell
#> 7.183      17     FALSE       7 17.82        1        1   table cell
#> 7.194      18     FALSE       7 17.42        1        1   table cell
#> 7.205      19     FALSE       7 19.47        1        1   table cell
#> 7.216      20     FALSE       7 18.52        1        1   table cell
#> 7.227      21     FALSE       7 19.90        1        1   table cell
#> 7.238      22     FALSE       7 20.01        1        1   table cell
#> 7.249      23     FALSE       7 16.87        1        1   table cell
#> 7.260      24     FALSE       7 17.30        1        1   table cell
#> 7.271      25     FALSE       7 15.41        1        1   table cell
#> 7.282      26     FALSE       7 17.05        1        1   table cell
#> 7.293      27     FALSE       7 18.90        1        1   table cell
#> 7.304      28     FALSE       7 16.70        1        1   table cell
#> 7.315      29     FALSE       7 16.90        1        1   table cell
#> 7.326      30     FALSE       7 14.50        1        1   table cell
#> 7.337      31     FALSE       7 15.50        1        1   table cell
#> 7.348      32     FALSE       7 14.60        1        1   table cell
#> 7.359      33     FALSE       7 18.60        1        1   table cell
#> 8.8         1      TRUE       8    vs        1        1   table cell
#> 8.19        2     FALSE       8     0        1        1   table cell
#> 8.30        3     FALSE       8     0        1        1   table cell
#> 8.41        4     FALSE       8     1        1        1   table cell
#> 8.52        5     FALSE       8     1        1        1   table cell
#> 8.63        6     FALSE       8     0        1        1   table cell
#> 8.74        7     FALSE       8     1        1        1   table cell
#> 8.85        8     FALSE       8     0        1        1   table cell
#> 8.96        9     FALSE       8     1        1        1   table cell
#> 8.107      10     FALSE       8     1        1        1   table cell
#> 8.118      11     FALSE       8     1        1        1   table cell
#> 8.129      12     FALSE       8     1        1        1   table cell
#> 8.140      13     FALSE       8     0        1        1   table cell
#> 8.151      14     FALSE       8     0        1        1   table cell
#> 8.162      15     FALSE       8     0        1        1   table cell
#> 8.173      16     FALSE       8     0        1        1   table cell
#> 8.184      17     FALSE       8     0        1        1   table cell
#> 8.195      18     FALSE       8     0        1        1   table cell
#> 8.206      19     FALSE       8     1        1        1   table cell
#> 8.217      20     FALSE       8     1        1        1   table cell
#> 8.228      21     FALSE       8     1        1        1   table cell
#> 8.239      22     FALSE       8     1        1        1   table cell
#> 8.250      23     FALSE       8     0        1        1   table cell
#> 8.261      24     FALSE       8     0        1        1   table cell
#> 8.272      25     FALSE       8     0        1        1   table cell
#> 8.283      26     FALSE       8     0        1        1   table cell
#> 8.294      27     FALSE       8     1        1        1   table cell
#> 8.305      28     FALSE       8     0        1        1   table cell
#> 8.316      29     FALSE       8     1        1        1   table cell
#> 8.327      30     FALSE       8     0        1        1   table cell
#> 8.338      31     FALSE       8     0        1        1   table cell
#> 8.349      32     FALSE       8     0        1        1   table cell
#> 8.360      33     FALSE       8     1        1        1   table cell
#> 9.9         1      TRUE       9    am        1        1   table cell
#> 9.20        2     FALSE       9     1        1        1   table cell
#> 9.31        3     FALSE       9     1        1        1   table cell
#> 9.42        4     FALSE       9     1        1        1   table cell
#> 9.53        5     FALSE       9     0        1        1   table cell
#> 9.64        6     FALSE       9     0        1        1   table cell
#> 9.75        7     FALSE       9     0        1        1   table cell
#> 9.86        8     FALSE       9     0        1        1   table cell
#> 9.97        9     FALSE       9     0        1        1   table cell
#> 9.108      10     FALSE       9     0        1        1   table cell
#> 9.119      11     FALSE       9     0        1        1   table cell
#> 9.130      12     FALSE       9     0        1        1   table cell
#> 9.141      13     FALSE       9     0        1        1   table cell
#> 9.152      14     FALSE       9     0        1        1   table cell
#> 9.163      15     FALSE       9     0        1        1   table cell
#> 9.174      16     FALSE       9     0        1        1   table cell
#> 9.185      17     FALSE       9     0        1        1   table cell
#> 9.196      18     FALSE       9     0        1        1   table cell
#> 9.207      19     FALSE       9     1        1        1   table cell
#> 9.218      20     FALSE       9     1        1        1   table cell
#> 9.229      21     FALSE       9     1        1        1   table cell
#> 9.240      22     FALSE       9     0        1        1   table cell
#> 9.251      23     FALSE       9     0        1        1   table cell
#> 9.262      24     FALSE       9     0        1        1   table cell
#> 9.273      25     FALSE       9     0        1        1   table cell
#> 9.284      26     FALSE       9     0        1        1   table cell
#> 9.295      27     FALSE       9     1        1        1   table cell
#> 9.306      28     FALSE       9     1        1        1   table cell
#> 9.317      29     FALSE       9     1        1        1   table cell
#> 9.328      30     FALSE       9     1        1        1   table cell
#> 9.339      31     FALSE       9     1        1        1   table cell
#> 9.350      32     FALSE       9     1        1        1   table cell
#> 9.361      33     FALSE       9     1        1        1   table cell
#> 10.10       1      TRUE      10  gear        1        1   table cell
#> 10.21       2     FALSE      10     4        1        1   table cell
#> 10.32       3     FALSE      10     4        1        1   table cell
#> 10.43       4     FALSE      10     4        1        1   table cell
#> 10.54       5     FALSE      10     3        1        1   table cell
#> 10.65       6     FALSE      10     3        1        1   table cell
#> 10.76       7     FALSE      10     3        1        1   table cell
#> 10.87       8     FALSE      10     3        1        1   table cell
#> 10.98       9     FALSE      10     4        1        1   table cell
#> 10.109     10     FALSE      10     4        1        1   table cell
#> 10.120     11     FALSE      10     4        1        1   table cell
#> 10.131     12     FALSE      10     4        1        1   table cell
#> 10.142     13     FALSE      10     3        1        1   table cell
#> 10.153     14     FALSE      10     3        1        1   table cell
#> 10.164     15     FALSE      10     3        1        1   table cell
#> 10.175     16     FALSE      10     3        1        1   table cell
#> 10.186     17     FALSE      10     3        1        1   table cell
#> 10.197     18     FALSE      10     3        1        1   table cell
#> 10.208     19     FALSE      10     4        1        1   table cell
#> 10.219     20     FALSE      10     4        1        1   table cell
#> 10.230     21     FALSE      10     4        1        1   table cell
#> 10.241     22     FALSE      10     3        1        1   table cell
#> 10.252     23     FALSE      10     3        1        1   table cell
#> 10.263     24     FALSE      10     3        1        1   table cell
#> 10.274     25     FALSE      10     3        1        1   table cell
#> 10.285     26     FALSE      10     3        1        1   table cell
#> 10.296     27     FALSE      10     4        1        1   table cell
#> 10.307     28     FALSE      10     5        1        1   table cell
#> 10.318     29     FALSE      10     5        1        1   table cell
#> 10.329     30     FALSE      10     5        1        1   table cell
#> 10.340     31     FALSE      10     5        1        1   table cell
#> 10.351     32     FALSE      10     5        1        1   table cell
#> 10.362     33     FALSE      10     4        1        1   table cell
#> 11.11       1      TRUE      11  carb        1        1   table cell
#> 11.22       2     FALSE      11     4        1        1   table cell
#> 11.33       3     FALSE      11     4        1        1   table cell
#> 11.44       4     FALSE      11     1        1        1   table cell
#> 11.55       5     FALSE      11     1        1        1   table cell
#> 11.66       6     FALSE      11     2        1        1   table cell
#> 11.77       7     FALSE      11     1        1        1   table cell
#> 11.88       8     FALSE      11     4        1        1   table cell
#> 11.99       9     FALSE      11     2        1        1   table cell
#> 11.110     10     FALSE      11     2        1        1   table cell
#> 11.121     11     FALSE      11     4        1        1   table cell
#> 11.132     12     FALSE      11     4        1        1   table cell
#> 11.143     13     FALSE      11     3        1        1   table cell
#> 11.154     14     FALSE      11     3        1        1   table cell
#> 11.165     15     FALSE      11     3        1        1   table cell
#> 11.176     16     FALSE      11     4        1        1   table cell
#> 11.187     17     FALSE      11     4        1        1   table cell
#> 11.198     18     FALSE      11     4        1        1   table cell
#> 11.209     19     FALSE      11     1        1        1   table cell
#> 11.220     20     FALSE      11     2        1        1   table cell
#> 11.231     21     FALSE      11     1        1        1   table cell
#> 11.242     22     FALSE      11     1        1        1   table cell
#> 11.253     23     FALSE      11     2        1        1   table cell
#> 11.264     24     FALSE      11     2        1        1   table cell
#> 11.275     25     FALSE      11     4        1        1   table cell
#> 11.286     26     FALSE      11     2        1        1   table cell
#> 11.297     27     FALSE      11     1        1        1   table cell
#> 11.308     28     FALSE      11     2        1        1   table cell
#> 11.319     29     FALSE      11     2        1        1   table cell
#> 11.330     30     FALSE      11     4        1        1   table cell
#> 11.341     31     FALSE      11     6        1        1   table cell
#> 11.352     32     FALSE      11     8        1        1   table cell
#> 11.363     33     FALSE      11     2        1        1   table cell